Test/requirements.txt
Requests==2.31.0
robotframework==6.1.1
